pyproject.toml
[project]
name = "dungeons-and-dragons"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
]
